✨ feat - Refreshes the new wallet creation page.
