+++ Source/Mobile/Finance/Controls/Buttons/IndicatorButton.xaml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <ContentView xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="Finance.Controls.IndicatorButtonError">
+             x:Class="Finance.Controls.Buttons.IndicatorButton">
 
     <Grid HeightRequest="44">
-        <Border x:Name="button" IsVisible="True" StrokeThickness="0" BackgroundColor="{AppThemeBinding Light={StaticResource Gray300}, Dark={StaticResource Gray600}}">
+        <Border x:Name="button" IsVisible="True" StrokeThickness="0" BackgroundColor="{AppThemeBinding Light={StaticResource Primary}, Dark={StaticResource PrimaryDark}}">
             <Border.StrokeShape>
                 <RoundRectangle CornerRadius="13" StrokeThickness="0" />
             </Border.StrokeShape>
@@ -14,8 +14,8 @@
             <Label VerticalTextAlignment="Center" HorizontalTextAlignment="Center">
                 <Label.FormattedText>
                     <FormattedString>
-                        <Span x:Name="icon" FontFamily="IconsSolid" FontSize="14" FontAttributes="Bold" TextColor="{AppThemeBinding Light={StaticResource Error}, Dark={StaticResource ErrorDark}}" />
-                        <Span x:Name="text" FontFamily="OpenSansRegular" FontSize="14" FontAttributes="Bold" TextColor="{AppThemeBinding Light={StaticResource Error}, Dark={StaticResource ErrorDark}}" />
+                        <Span x:Name="icon" FontFamily="IconsSolid" FontSize="14" FontAttributes="Bold" TextColor="{StaticResource Gray000}" />
+                        <Span x:Name="text" FontFamily="OpenSansRegular" FontSize="14" FontAttributes="Bold" TextColor="{StaticResource Gray000}" />
                     </FormattedString>
                 </Label.FormattedText>
                 <VisualStateManager.VisualStateGroups>
@@ -24,7 +24,7 @@
                             <VisualState x:Name="Normal" />
                             <VisualState x:Name="Disabled">
                                 <VisualState.Setters>
-                                    <Setter Property="TextColor" Value="{AppThemeBinding Light={StaticResource Gray600}, Dark={StaticResource Gray200}}" />
+                                    <Setter Property="TextColor" Value="{StaticResource Gray000}" />
                                 </VisualState.Setters>
                             </VisualState>
                         </VisualStateGroup>
@@ -37,18 +37,18 @@
                         <VisualState x:Name="Normal" />
                         <VisualState x:Name="Disabled">
                             <VisualState.Setters>
-                                <Setter Property="BackgroundColor" Value="{AppThemeBinding Light={StaticResource Gray200}, Dark={StaticResource Gray900}}" />
+                                <Setter Property="BackgroundColor" Value="{AppThemeBinding Light={StaticResource PrimaryDisabled}, Dark={StaticResource PrimaryDarkDisabled}}" />
                             </VisualState.Setters>
                         </VisualState>
                     </VisualStateGroup>
                 </VisualStateGroupList>
             </VisualStateManager.VisualStateGroups>
         </Border>
-        <Border x:Name="frame" IsVisible="False" StrokeThickness="0" BackgroundColor="{AppThemeBinding Light={StaticResource Gray300}, Dark={StaticResource Gray600}}">
+        <Border x:Name="frame" IsVisible="False" StrokeThickness="0" BackgroundColor="{AppThemeBinding Light={StaticResource Primary}, Dark={StaticResource PrimaryDark}}">
             <Border.StrokeShape>
                 <RoundRectangle CornerRadius="13" StrokeThickness="0" />
             </Border.StrokeShape>
-            <ActivityIndicator IsRunning="True" Color="{AppThemeBinding Light={StaticResource Gray600}, Dark={StaticResource Gray400}}" IsVisible="True" HeightRequest="25" />
+            <ActivityIndicator IsRunning="True" Color="{StaticResource Gray000}" IsVisible="True" HeightRequest="25" />
         </Border>
     </Grid>
 </ContentView>
