+++ Source/Mobile/Finance/Controls/Buttons/IndicatorButton.xaml.cs
@@ -1,12 +1,12 @@
 using System.Windows.Input;
 
-namespace Finance.Controls;
+namespace Finance.Controls.Buttons;
 
-public partial class IndicatorButtonError : ContentView {
-    public static readonly BindableProperty GlyphProperty = BindableProperty.Create(nameof(Glyph), typeof(string), typeof(IndicatorButtonError), string.Empty, propertyChanged: OnGlyphChanged);
-    public static readonly BindableProperty IsRunningProperty = BindableProperty.Create(nameof(IsRunning), typeof(bool), typeof(IndicatorButtonError), false, propertyChanged: OnIsRunningChanged);
-    public static readonly BindableProperty TextProperty = BindableProperty.Create(nameof(Text), typeof(string), typeof(IndicatorButtonError), string.Empty, propertyChanged: OnTextChanged);
-    public static readonly BindableProperty CommandProperty = BindableProperty.Create(nameof(Command), typeof(ICommand), typeof(IndicatorButtonError));
+public partial class IndicatorButton : ContentView {
+    public static readonly BindableProperty GlyphProperty = BindableProperty.Create(nameof(Glyph), typeof(string), typeof(IndicatorButton), string.Empty, propertyChanged: OnGlyphChanged);
+    public static readonly BindableProperty IsRunningProperty = BindableProperty.Create(nameof(IsRunning), typeof(bool), typeof(IndicatorButton), false, propertyChanged: OnIsRunningChanged);
+    public static readonly BindableProperty TextProperty = BindableProperty.Create(nameof(Text), typeof(string), typeof(IndicatorButton), string.Empty, propertyChanged: OnTextChanged);
+    public static readonly BindableProperty CommandProperty = BindableProperty.Create(nameof(Command), typeof(ICommand), typeof(IndicatorButton));
 
     public string Glyph {
         get { return (string)GetValue(GlyphProperty); }
@@ -30,23 +30,23 @@ public partial class IndicatorButtonError : ContentView {
 
     public event EventHandler Clicked;
 
-    public IndicatorButtonError() {
+    public IndicatorButton() {
         InitializeComponent();
     }
 
     private static void OnGlyphChanged(BindableObject bindable, object oldValue, object newValue) {
-        if(bindable is not IndicatorButtonError control) { return; }
+        if(bindable is not IndicatorButton control) { return; }
         control.icon.Text = $"{newValue}   ";
     }
 
     private static void OnIsRunningChanged(BindableObject bindable, object oldValue, object newValue) {
-        if(bindable is not IndicatorButtonError control) { return; }
+        if(bindable is not IndicatorButton control) { return; }
         control.frame.IsVisible = (bool)newValue;
         control.button.IsVisible = !(bool)newValue;
     }
 
     private static void OnTextChanged(BindableObject bindable, object oldValue, object newValue) {
-        if(bindable is not IndicatorButtonError control) { return; }
+        if(bindable is not IndicatorButton control) { return; }
         control.text.Text = newValue.ToString();
     }
