+++ Source/Mobile/Finance/Controls/Navigation/NavigationBar.xaml
@@ -2,7 +2,7 @@
 <ContentView xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:controls_devexpress="clr-namespace:DevExpress.Maui.Controls;assembly=DevExpress.Maui.Controls"
-             x:Class="Finance.Controls.NavigationBar"
+             x:Class="Finance.Controls.Navigation.NavigationBar"
              NavigationPage.HasNavigationBar="False">
 
     <ContentView.Content>
