+++ Source/Mobile/Finance/Controls/Navigation/ToolbarItem.xaml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <ContentView xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="Finance.Controls.ToolbarItem">
+             x:Class="Finance.Controls.Navigation.ToolbarItem">
 
     <Border HeightRequest="40" WidthRequest="40" StrokeThickness="0">
         <Border.StrokeShape>
