+++ Source/Mobile/Finance/Finance.csproj.user
@@ -17,19 +17,19 @@
     <MauiXaml Update="Controls\Cards\AssetCard.xaml">
       <SubType>Designer</SubType>
     </MauiXaml>
-    <MauiXaml Update="Controls\IndicatorButton.xaml">
+    <MauiXaml Update="Controls\Buttons\IndicatorButton.xaml">
       <SubType>Designer</SubType>
     </MauiXaml>
-    <MauiXaml Update="Controls\IndicatorButtonError.xaml">
+    <MauiXaml Update="Controls\Buttons\IndicatorButtonError.xaml">
       <SubType>Designer</SubType>
     </MauiXaml>
-    <MauiXaml Update="Controls\NavigationBar.xaml">
+    <MauiXaml Update="Controls\Navigation\NavigationBar.xaml">
       <SubType>Designer</SubType>
     </MauiXaml>
-    <MauiXaml Update="Controls\SmallButtonCard.xaml">
+    <MauiXaml Update="Controls\Buttons\SmallButtonCard.xaml">
       <SubType>Designer</SubType>
     </MauiXaml>
-    <MauiXaml Update="Controls\ToolbarItem.xaml">
+    <MauiXaml Update="Controls\Navigation\ToolbarItem.xaml">
       <SubType>Designer</SubType>
     </MauiXaml>
     <MauiXaml Update="Pages\DangerZone\DangerZonePage.xaml">
@@ -38,7 +38,7 @@
     <MauiXaml Update="Pages\DangerZone\Delete\DeleteWalletPage.xaml">
       <SubType>Designer</SubType>
     </MauiXaml>
-    <MauiXaml Update="Pages\CreateWallet\CreateWalletPage.xaml">
+    <MauiXaml Update="Pages\Walleting\CreateWallet\CreateWalletPage.xaml">
       <SubType>Designer</SubType>
     </MauiXaml>
     <MauiXaml Update="Pages\Historic\HistoricPage.xaml">
