+++ Source/Mobile/Finance/GlobalSuppressions.cs
@@ -1,0 +1,3 @@
+﻿using System.Diagnostics.CodeAnalysis;
+
+[assembly: SuppressMessage("Toolkit", "MVVMTK0045", Justification = "AOT compatibility not required for this MAUI app targeting Windows", Scope = "module")]
