+++ Source/Mobile/Finance/Pages/DangerZone/Delete/DeleteWalletPage.xaml
@@ -1,7 +1,8 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:controls="clr-namespace:Finance.Controls"
+             xmlns:buttons="clr-namespace:Finance.Controls.Buttons"
+             xmlns:navigation="clr-namespace:Finance.Controls.Navigation"
              xmlns:editors="clr-namespace:DevExpress.Maui.Editors;assembly=DevExpress.Maui.Editors"
              xmlns:styles="clr-namespace:Finance.Resources.Styles"
              x:Class="Finance.Pages.DangerZone.Delete.DeleteWalletPage"
@@ -9,7 +10,7 @@
              Shell.FlyoutBehavior="Disabled">
 
     <Grid RowDefinitions="Auto,*,Auto,Auto">
-        <controls:NavigationBar Title="Deletar Carteira" IsRoot="False" Clicked="BackButtom_Clicked" />
+        <navigation:NavigationBar Title="Deletar Carteira" IsRoot="False" Clicked="BackButtom_Clicked" />
 
         <VerticalStackLayout Grid.Row="1" Margin="15,0" Spacing="10">
             <Label FontSize="20">
@@ -32,6 +33,6 @@
             </Label>
         </HorizontalStackLayout>
 
-        <controls:IndicatorButton Grid.Row="3" Margin="20,0,20,30" Glyph="&#xf2ed;" IsEnabled="{Binding Source={x:Reference checkBox}, Path=IsChecked}" IsRunning="{Binding IsRunning}" Text="Apagar Carteira" Command="{Binding DeleteCommand}" />
+        <buttons:IndicatorButton Grid.Row="3" Margin="20,0,20,30" Glyph="&#xf2ed;" IsEnabled="{Binding Source={x:Reference checkBox}, Path=IsChecked}" IsRunning="{Binding IsRunning}" Text="Apagar Carteira" Command="{Binding DeleteCommand}" />
     </Grid>
 </ContentPage>
