+++ Source/Mobile/Finance/Pages/LoadingViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using Finance.Services;
+using Finance.Services.Walleting;
 
 namespace Finance.Pages;
