+++ Source/Mobile/Finance/Pages/SelectWallet/SelectWalletPage.xaml
@@ -2,19 +2,19 @@
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:collection="clr-namespace:DevExpress.Maui.CollectionView;assembly=DevExpress.Maui.CollectionView"
-             xmlns:controls="clr-namespace:Finance.Controls"
+             xmlns:navigation="clr-namespace:Finance.Controls.Navigation"
              x:Class="Finance.Pages.SelectWallet.SelectWalletPage"
              Shell.NavBarIsVisible="False"
              Shell.FlyoutBehavior="Disabled">
 
     <Grid RowDefinitions="Auto,*">
-        <controls:NavigationBar Title="Selecionar Carteira" IsRoot="False" Clicked="BackButtom_Clicked">
-            <controls:NavigationBar.ToolbarItems>
+        <navigation:NavigationBar Title="Selecionar Carteira" IsRoot="False" Clicked="BackButtom_Clicked">
+            <navigation:NavigationBar.ToolbarItems>
                 <VerticalStackLayout>
-                    <controls:ToolbarItem Glyph="&#x2b;" Command="{Binding NewWalletCommand}" />
+                    <navigation:ToolbarItem Glyph="&#x2b;" Command="{Binding NewWalletCommand}" />
                 </VerticalStackLayout>
-            </controls:NavigationBar.ToolbarItems>
-        </controls:NavigationBar>
+            </navigation:NavigationBar.ToolbarItems>
+        </navigation:NavigationBar>
         <Grid Grid.Row="1">
             <Border Grid.Row="2" Grid.ColumnSpan="3" IsVisible="{Binding IsEmpty}" HeightRequest="150" Style="{x:StaticResource Card}" Margin="15,0" VerticalOptions="Start">
                 <Label Text="Não há outra carteira para selecionar aqui!&#10;&#10;Você pode criar uma nova carteira no botão + no canto superior direito da tela." FontSize="Micro" TextColor="Gray" VerticalTextAlignment="Center" HorizontalTextAlignment="Center" Padding="20" />
